# Film Manager Implementation/api/openapi.yaml
openapi: 3.0.1
info:
  title: Film Manager
  version: 1.0.0
  description: REST APIs that alllows to interact with the Film Manager Service
...
paths:
...
  "/api/films":
    post:
      tags: 
      - "api/films"
      summary: Used to add a new film to the owner list
      description: An authenticated user can create a film that is added to it's own list
      security:
        - cookieAuth: []
      operationId: createFilm
      requestBody:
        description: Represents the film that is going to be created
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
            examples:
              example1:
                value:
                  title: Inside Out 2
                  owner: 1
                  private: true
                  watchDate: "2024-11-27"
                  rating : 9
                  favorite: false
              example2:
                value:
                  title: Inception
                  owner: 1
                  private: false
        required: true
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
              examples:
                example1:
                  value:
                    id: 10
                    title: Inside Out 2
                    owner: 1
                    private: true
                    watchDate: "2024-11-27"
                    rating : 9
                    favorite: false
                    self: /api/films/public/10
                example2:
                  value:
                    id: 11
                    title: Inception,
                    owner: 1
                    private: false
                    reviews: /api/films/public/11/reviews
                    self: /api/films/public/11
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
    get:
      tags: 
      - "api/films"
      summary: Retrive all the films created by the authenticated user.
      description: An authenticated user can create all the films that he created.  A pagination machanism is used in order to limit the size of the messages.
      security:
        - cookieAuth: []
      operationId: getOwnedFilms
      parameters:
        - in: query
          name: pageNo                  
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example:
                {totalPages: 3,
                currentPage: 1,
                films: [
                    {
                      id: 10,
                      title: Inside Out 2,
                      owner: 1,
                      private: true,
                      watchDate: "2024-11-27",
                      rating : 9,
                      favorite: false,
                      self: /api/films/public/10
                    },
                    {
                        id: 11,
                        title: Inception,
                        owner: 1,
                        private: false,
                        reviews: /api/films/public/11/reviews,
                        self: /api/films/public/11
                    }
                ],
                next: /api/films?pageNo=2
                }
            


        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  "/api/films/private":
    get:
      tags: 
      - "api/films/private"
      security:
        - cookieAuth: []
      operationId: getPrivateFilms
      summary: retrieve the private films of the user.
      description: A logged-in user can retrieve his private films. A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requeted page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 1,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 1,
                          "title": "Your Name",
                          "owner": 1,
                          "private": true,
                          "watchDate": "2021-10-03",
                          "rating": 9,
                          "self": "/api/films/private/1"
                      },
                      {
                          "id": 4,
                          "title": "Weathering with You",
                          "owner": 1,
                          "private": true,
                          "self": "/api/films/private/4"
                      }
                  ],
                  "next": null
              }
              
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  "/api/films/private/{filmId}":
    get:
      tags: 
      - "api/films/private/{filmId}"
      security:
        - cookieAuth: []
      operationId: getSinglePrivateFilm
      summary: retrieve a specific private film of the user.
      description: A logged-in user can retrieve a specific private film that he owns.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
              example: 
                id: 10
                title: Inside Out 2
                owner: 1
                private: true
                watchDate: "2024-11-27"
                rating : 9
                favorite: false
                self: /api/films/public/10

        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    put:
      tags: 
      - "api/films/private/{filmId}"
      security:
        - cookieAuth: []
      operationId: updateSinglePrivateFilm
      summary: Update a specific private film of the user.
      description: A logged-in user can update a specific private film that he owns except for it's visibility.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      requestBody:
        description: The film object that must be replaced to the old one
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
            example:
              title: Inside Out 2
              owner: 1
              private: true
              watchDate: "2024-11-27"
              rating : 10
              favorite: true
        required: true
      responses:
        "200":
          description: Succesfully Updated           
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (The request is not allowed, like a change of visibility)
        "500":
          description: Internal Server Error

    delete:
      tags: 
      - "api/films/private/{filmId}"
      security:
        - cookieAuth: []
      operationId: deleteSinglePrivateFilm
      summary: Delete a specific private film of the user.
      description: A logged-in user can delete a specific private film that he owns.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 10
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
  "/api/films/public":
    get:
      tags: 
      - "api/films/public"
      operationId: getPublicFilms
      summary: retrieve a list of public films.
      description: A generic user can retrieve a list of public films added by other users. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 3,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 2,
                          "title": "Heaven's Feel",
                          "owner": 2,
                          "private": false,
                          "reviews": "/api/films/public/2/reviews",
                          "self": "/api/films/public/2"
                      },
                      {
                          "id": 3,
                          "title": "You Can (Not) Redo",
                          "owner": 1,
                          "private": false,
                          "reviews": "/api/films/public/3/reviews",
                          "self": "/api/films/public/3"
                      }
                  ],
                  "next": "/api/films/public?pageNo=2"
              }
        "500":
          description: Internal Server Error
  "/api/films/public/invited":
    get:
      security:
      - cookieAuth: []
      tags:
      - api/films/public/invited
      summary: Retrieve the public films that the logged-user has been invited to review
      description: The logged-in user can retrieve all the public films he was invited to review. 
        A pagination machanism is used in order to limit the size of the messages.
      operationId: getInvitedFilms
      parameters:
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Films"
              example: 
                {
                  "totalPages": 2,
                  "currentPage": 1,
                  "films": [
                      {
                          "id": 3,
                          "title": "You Can (Not) Redo",
                          "owner": 1,
                          "private": false,
                          "reviews": "/api/films/public/3/reviews",
                          "self": "/api/films/public/3"
                      },
                      {
                          "id": 4,
                          "title": "Weathering with You",
                          "owner": 1,
                          "private": false,
                          "reviews": "/api/films/public/4/reviews",
                          "self": "/api/films/public/4"
                      }
                  ],
                  "next": "/api/films/public/invited?pageNo=2"
              }
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
  "/api/films/public/{filmId}":
    get:
      tags: 
      - "api/films/public/{filmId}"
      operationId: getSinglePublicFilm
      summary: retrieve a specific public film.
      description: A generic user can retrieve a specific public film.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Film"
              example: 
                id: 11
                title: Inception
                owner: 1
                private: false
                reviews: /api/films/public/11/reviews
                self: /api/films/public/11
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    put:
      tags: 
      - "api/films/public/{filmId}"
      security:
        - cookieAuth: []
      operationId: updateSinglePublicFilm
      summary: Update a specific public film of the user.
      description: A logged-in user can update a specific public film that he owns except for it's visibility.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      requestBody:
        description: The film object that must be replaced to the old one
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Film"
            example: 
              title: Inception Remake
              owner: 1
              private: false
        required: true
      responses:
        "200":
          description: Succesfully Updated
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (The request is not allowed, like a change of visibility)
        "500":
          description: Internal Server Error

    delete:
      tags: 
      - "api/films/public/{filmId}"
      security:
        - cookieAuth: []
      operationId: deleteSinglePublicFilm
      summary: Delete a specific public film of the user.
      description: A logged-in user can delete a specific public film that he owns.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 11
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
  "/api/films/public/{filmId}/reviews":
    get:
      tags: 
      - "api/films/public/{filmId}/reviews"
      operationId: getReviews
      summary: retrieve the reviews of a specific public film.
      description: A generic user can retrieve all the reviews of a specific public film. 
        A pagination machanism is used in order to limit the size of the messages.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - in: query
          name: pageNo
          required: false
          schema:
            type: integer
            default: 1
          description: The number of the requested page. It's optional so if not expessed the first page is returned.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Reviews"
              example: 
                {
                  "totalPages": 1,
                  "currentPage": 1,
                  "reviews": [
                      {
                          "filmId": 2,
                          "reviewerId": 2,
                          "completed": true,
                          "reviewDate": "2022-04-04",
                          "rating": 10,
                          "review": "I could not stop watching until the very end. However, you must watch the prequels before this film to really understand it.",
                          "self": "/api/films/public/2/reviews/2"
                      },
                      {
                          "filmId": 2,
                          "reviewerId": 5,
                          "completed": true,
                          "reviewDate": "2022-03-04",
                          "rating": 10,
                          "review": "This film is a perfect conclusion for the trilogy.",
                          "self": "/api/films/public/2/reviews/5"
                      }
                  ],
                  "next": null
              }
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    post:
      tags: 
      - "api/films/public/{filmId}/reviews"
      security:
      - cookieAuth: []
      operationId: issueFilmReview
      summary: Issue film review to some users
      description: The owner of the public film, identified by filmId, assign it to one or more users and the relative reviews are created
                   with the completed property set to false.
      parameters:
      - name: filmId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example : 10
      requestBody:
        description: An array of reviewers . Each item in the array identifies a user to whom 
          the review for a specific film (identified by the filmId parameter) should be issued.
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  reviewerId:
                    type: integer
              example: 
                [{
                    "reviewerId":5
                  },{
                    "reviewerId":4                 
                  }] 
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                  "$ref": "#/components/schemas/Reviews"
              example: 
                [
                  {
                      "filmId": 4,
                      "reviewerId": 4,
                      "completed": false,
                      "self": "/api/films/public/4/reviews/4"
                  },
                  {
                      "filmId": 4,
                      "reviewerId": 5,
                      "completed": false,
                      "self": "/api/films/public/4/reviews/5"
                  }
              ]
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (user is not the owner)
        '404':
          description: Not Found
        '409':
          description: Conflict (e.g., wrong film id or unknown reviewer id)
        '500':
          description: Internal Server Error      
  "/api/films/public/{filmId}/reviews/{reviewerId}":
    get:
      tags: 
      - "api/films/public/{filmId}/reviews/{reviewerId}"
      operationId: getSingleReview
      summary: retrieve a specific review of a public film.
      description: A generic user can retrieve a specific completed review of a public film.
        An authenticated user can retrive an incomplete review if it's assigned to him 
        or if he is the owner of the film
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 2
        - name: reviewerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 5
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Review"
              examples:
                example1:
                  value:
                    filmId: 2
                    reviewerId: 5
                    completed: true
                    reviewDate: "2022-03-04"
                    rating": 10
                    review: This film is a perfect conclusion for the trilogy.
                    self: "/api/films/public/2/reviews/5"
              
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    put:
      tags: 
      - "api/films/public/{filmId}/reviews/{reviewerId}"
      security:
      - cookieAuth: []
      operationId: completeReview
      summary: Complete a review
      description: This operation allows to set the review, identified by the pair filmId, reviewerId, as completed.
         The completed property is set to true and the reviewDate, rating and review properties are changed
         according to the values specified by the user that completed the review.
      parameters:
      - name: filmId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 4
      - name: reviewerId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      requestBody:
        description: The completed review 
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Review"
            example: 
              {
                "filmId": 4,
                "reviewerId": 1,
                "completed": true,
                "reviewDate": "2025-01-04",
                "rating": 9,
                "review": "Amazing story!",
              }

                
        required: true
            
      responses:
        "204":
          description: Succesfully completed
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
    delete:
      tags: 
      - "api/films/public/{filmId}/reviews/{reviewerId}"
      security:
      - cookieAuth: []
      operationId: deleteReviewInvitation
      summary: Delete a review invitation
      description: This operation can be performed only by the film's owner on a review that is not yet completed.
      parameters:
        - name: filmId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 6
        - name: reviewerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
            example: 1
      responses:
        "204":
          description: Successfully Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not Found
        "409":
          description: Conflict (e.g. the review is already completed)
        "500":
          description: Internal Server Error
...
  "/api/users":
    get:
      tags: 
      - "api/users"
      security:
      - cookieAuth: []
      summary: Get informations about users
      description: An authenticated user can retrieve informations about all the other users (Password excluded)
      operationId: getUsers
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Users"
              example: 
                [
                  {
                      "id": 1,
                      "name": "User",
                      "email": "user.dsp@polito.it",
                      "self": "/api/users/1"
                  },
                  {
                      "id": 2,
                      "name": "Frank",
                      "email": "frank.stein@polito.it",
                      "self": "/api/users/2"
                  },
                  {
                      "id": 3,
                      "name": "Karen",
                      "email": "karen.makise@polito.it",
                      "self": "/api/users/3"
                  },
                  {
                      "id": 4,
                      "name": "Rene",
                      "email": "rene.regeay@polito.it",
                      "self": "/api/users/4"
                  },
                  {
                      "id": 5,
                      "name": "Beatrice",
                      "email": "beatrice.golden@polito.it",
                      "self": "/api/users/5"
                  },
                  {
                      "id": 6,
                      "name": "Arthur",
                      "email": "arthur.pendragon@polito.it",
                      "self": "/api/users/6"
                  }
                  ]
        '401':
          description: Unauthorized
        '404':
          description: Not Found
  "/api/users/{userId}":
    get:
      tags: 
      - "api/users/{userId}"
      security:
      - cookieAuth: []
      summary: Get informations about a specific user 
      description: An authenticated user can retrieve informations about a user idetified by userId (Password excluded)
      operationId: getSingleUser
      parameters:
      - name: userId
        in: path
        description: unique identifier of the user to get
        required: true
        schema:
          type: integer
          format: int64
          example: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/User"
              example: 
                {
                    "id": 1,
                    "name": "User",
                    "email": "user.dsp@polito.it",
                    "self": "/api/users/1"
                }
        '401':
          description: Unauthorized
        '404':
          description: Not Found
...
  "/api/films/public/reviews/modifications/{mId}/accept":
    put: 
      tags:
        - api/film/public/reviews/modifications/{mId}/accept
      summary: Used to accept a review modification request
      description: The owner of the film can accept a pending review modification request
      security:
        - cookieAuth: []
      operationId: acceptReviewModificationRequest
      parameters:
      - name: mId
        in: path
        required: true
        schema:
          type: integer
          format: int64
          example: 3
      responses:
        "204":
          description: Succesfully accepted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "409":
          description: Conflict (The modification request is not pending)
        "500":
          description: Internal Server Error    
...
components:
  schemas:
...
    Users:
      type: array
      items:
        "$ref": "#/components/schemas/User"
      description: Array of User objects representing the list of users.
    Film:
      type: object
      description: The schema for the film data structure
      additionalProperties: false
      properties:
        "$schema":
          type: string
          description: The path to the schema that describes the film data structure
        id:
          type: integer
          description: The film's unique identifier
        title:
          type: string
          description: The film's title
        owner:
          type: integer
          description: The id of the user that created the film data structure
        private:
          type: boolean
          default: true
          description: It's true if the owner is the only one that can access the film, false otherwise
        watchDate:
          type: string
          format: date
          description: The date when the owner watched the Film
        rating:
          type: integer
          minimum: 0
          maximum: 10
          description: the rating that the owner has given to the film
        favourite:
          type: boolean
          default: false
          description: It's true if the film is among the owner favourites, false otherwise
        reviews:
          type: string
          format: uri
          description: the reviews that other users have given to the film
        self:
          type: string
          format: uri
          description: URI to access the specific film resource.
      required:
        - title
        - owner
        - private
...
  securitySchemes:
    cookieAuth:
      description: JWT to identify the authenticated user.
      type: apiKey
      in: cookie
      name: connect.sid